Move cli-related files to subfolder
