+++ main/cli/cli-commands-device-info.c
